[Feat] #50 - 파일 분리 및 주석 추가
